+++ changelog.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.0.0-alpha.1] - Not Yet Released
+## [3.0.0-alpha.2] - Not Yet Released
+
+### Fixed
+* Recursive unions are now supported (issue [#54](https://github.com/JasonBock/CslaGeneratorSerialization/issues/54))
+
+## [3.0.0-alpha.1] - 2026-08-11
 
 ### Added
 * Added support for union types (issue [#49](https://github.com/JasonBock/CslaGeneratorSerialization/issues/49))